backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
django-filter>=24.0
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
